refactor: move data logic to separate mutator classes
Removed 'not remove all if numerically indexed' test / code because this was faulty behavior.
